backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.50.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
python-docx>=1.1.0
reportlab>=4.0.0
pydantic-settings>=2.0.0
